+++ Frontend/src/lib/components/index.ts
@@ -1,5 +1,6 @@
 export {default as Header} from './Header.svelte';
 export {default as Footer} from './Footer.svelte';
+export {default as Slogan} from './Slogan.svelte';
 export {default as Notification} from './Notification.svelte';
 
 export {default as CheveronImg} from './icons/Cheveron.svelte';